backend/pyproject.toml
[project]
name = "artisan-market-backend"
version = "0.1.0"
description = "Async FastAPI backend for artisan marketplace"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "asyncpg>=0.29.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
]

[tool.uv]
package = false
